+++ lumin/optimisation/features.py
@@ -13,7 +13,16 @@ from ..utils.mod_ver import check_rfpimp
 
 
 def get_rf_feat_importance(rf:ForestRegressor, inputs:pd.DataFrame, targets:np.ndarray, weights:Optional[np.ndarray]=None) -> pd.DataFrame:
-    '''Wrapper function for rfpimp which checks correct version is installed'''
+    r'''
+    Compute feature importance for a Random Forest model using rfpimp.
+    This requires a bleeding edge installation of rfpimp, which is not available from PIP.
+
+    Arguments:
+        rf: trained Random Forest model
+        inputs: input data as Pandas DataFrame
+        targets: target data as Numpy array
+        weights: Optional data weights as Numpy array
+    '''
     check_rfpimp(); from rfpimp import importances
     return importances(rf, inputs, targets, features=inputs.columns, sample_weights=weights).reset_index()
 
@@ -21,8 +30,31 @@ def get_rf_feat_importance(rf:ForestRegressor, inputs:pd.DataFrame, targets:np.n
 def rf_rank_features(train_df:pd.DataFrame, val_df:pd.DataFrame, objective:str,
                      train_feats:List[str], targ_name:str='gen_target', wgt_name:Optional[str]=None,
                      importance_cut:float=0.0, n_estimators:int=40, n_rfs:int=1,
-                     savename:Optional[str]=None, settings:PlotSettings=PlotSettings()) -> List[str]:
-    '''Rank train_feats via permutation importance using random forests'''
+                     savename:Optional[str]=None, plot_settings:PlotSettings=PlotSettings()) -> List[str]:
+    r'''
+    Compute relative permutation importance of input features via using Random Forests.
+    A reduced set of 'important features' is obtained by cutting on relative importance and a new model is trained and evaluated on this reduced set.
+    RFs will have their hyper-parameters roughly optimised, both when training on all features and once when training on important features.
+    Relative importances may be computed multiple times (via n_rfs) and averaged. In which case the standard error is also computed.
+    Currently, this requires a bleeding edge installation of rfpimp, which is not available from PIP.
+
+    Arguments:
+        train_df: training data as Pandas DataFrame
+        val_df: validation data as Pandas DataFrame
+        objective: string representation of objective: either 'classification' or 'regression'
+        train_feats: complete list of training features
+        targ_name: name of column containing target data
+        wgt_name: name of column containing weight data. If set, will use weights for training and evaluation, otherwise will not
+        importance_cut: minimum importance required to be considered an 'important feature'
+        n_estimators: number of trees to use in each forest
+        n_rfs: number of trainings to perform on all training features in order to compute importances
+        savename: Optional name of file to which to save the plot of feature importances
+        plot_settings: :class:PlotSettings class to control figure appearance
+
+    Returns:
+        List of features passing importance_cut, ordered by importance
+    '''
+
     w_trn = None if wgt_name is None else train_df[wgt_name]
     w_val = None if wgt_name is None else val_df[wgt_name]
     print("Optimising RF")
@@ -43,7 +75,7 @@ def rf_rank_features(train_df:pd.DataFrame, val_df:pd.DataFrame, objective:str,
         orig_score /= n_rfs
         fi.sort_values(by='Importance', ascending=False, inplace=True)
     print("Top ten most important features:\n", fi[['Feature', 'Importance']][:min(len(fi), 10)])
-    plot_importance(fi[:min(len(fi), 30)], savename=savename, settings=settings)
+    plot_importance(fi[:min(len(fi), 30)], savename=savename, settings=plot_settings)
 
     top_feats = list(fi[fi.Importance > importance_cut].Feature)
     print(f"\n{len(top_feats)} features found with importance greater than {importance_cut}:\n", top_feats, '\n')